+++ src/index.ts
@@ -1,1 +1,2 @@
 export { Ikat } from './ikat';
+export { IkatV2 } from './ikat-v2';